[refactor] Moved carbon api caller to its own file: better semantics.
