+++ lib/widgets/electricty.dart
@@ -2,6 +2,7 @@ import 'package:electricity_prices_and_carbon_intensity/utilities/electricityApi
 import 'package:electricity_prices_and_carbon_intensity/utilities/httpclient.dart';
 import 'package:electricity_prices_and_carbon_intensity/utilities/style.dart';
 import 'package:electricity_prices_and_carbon_intensity/widgets/chart.dart';
+import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
